remove module station to favorite station and create private method in stations from wqp
